+++ index.html
@@ -227,6 +227,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-
+<script type = 'module' src = './js/main.js'></script>
 </body>
 </html>
